added deeplinks to its own file
