+++ aries-site/src/tests/navigation/homepage.js
@@ -28,17 +28,3 @@ test('should navigate to correct path when a tile is choosen via keyboard', asyn
     .contains(expectedPath);
 });
 
-/*Added range to this test since now there is some issues with different browsers as to where
-the page actually lands when routed to a deep link
-*/
-
-fixture('Deep linking').page(baseUrl);
-
-test('should navigate to correct hash of page when a deep link is directly routed to', async t => {
-  const url = `${baseUrl}/foundation/color#background-colors`
-  const pageSection = Selector('#background-colors');
-  await t
-    .navigateTo(url)
-    .expect(pageSection.getBoundingClientRectProperty("top"))
-    .within(-150, 150);
-});
